refactor: Winners와 WinnersDto 클래스 분리
