+++ src/main/kotlin/racingcar/controller/RacingController.kt
@@ -1,7 +1,7 @@
 package racingcar.controller
 
 import racingcar.dto.car.CarsDto
-import racingcar.model.car.WinnersDto
+import racingcar.dto.car.WinnersDto
 import racingcar.model.round.RoundDto
 import racingcar.service.RacingService
 import racingcar.view.InputView
