+++ src/main/kotlin/racingcar/dto/car/WinnersDto.kt
@@ -1,0 +1,5 @@
+package racingcar.dto.car
+
+import racingcar.model.car.Car
+
+class WinnersDto(winners: List<Car>) : List<Car> by winners
