+++ src/main/kotlin/racingcar/model/car/Winners.kt
@@ -1,7 +1,7 @@
 package racingcar.model.car
 
+import racingcar.dto.car.WinnersDto
+
 class Winners(winners: List<Car>) : List<Car> by winners {
     fun toDto(): WinnersDto = WinnersDto(this)
 }
-
-class WinnersDto(winners: List<Car>) : List<Car> by winners
